feat: transforming no grant block into component, select closes dropdown
