+++ utilities/pages.ts
@@ -32,6 +32,7 @@ export const PAGES = {
     TABS: {
       OVERVIEW: (project: string, grant: string) =>
         `/project/${project}/grants?grantId=${grant}&tab=overview`,
+
       SELECTED_TAB: (project: string, grant?: string, tab?: string) =>
         `/project/${project}/grants?${grant ? `&grantId=${grant}` : ""}${
           tab ? `&tab=${tab}` : ""
